+++ src/thunderbots/software/ai/main.cpp
@@ -76,14 +76,14 @@ int main(int argc, char **argv)
 
     // Create subscribers
     ros::Subscriber field_sub = node_handle.subscribe(
-        Util::Constants::BACKEND_INPUT_FIELD_TOPIC, 1, fieldUpdateCallback);
+        Util::Constants::NETWORK_INPUT_FIELD_TOPIC, 1, fieldUpdateCallback);
     ros::Subscriber ball_sub = node_handle.subscribe(
-        Util::Constants::BACKEND_INPUT_BALL_TOPIC, 1, ballUpdateCallback);
+        Util::Constants::NETWORK_INPUT_BALL_TOPIC, 1, ballUpdateCallback);
     ros::Subscriber friendly_team_sub =
-        node_handle.subscribe(Util::Constants::BACKEND_INPUT_FRIENDLY_TEAM_TOPIC, 1,
+        node_handle.subscribe(Util::Constants::NETWORK_INPUT_FRIENDLY_TEAM_TOPIC, 1,
                               friendlyTeamUpdateCallback);
     ros::Subscriber enemy_team_sub = node_handle.subscribe(
-        Util::Constants::BACKEND_INPUT_ENEMY_TEAM_TOPIC, 1, enemyTeamUpdateCallback);
+        Util::Constants::NETWORK_INPUT_ENEMY_TEAM_TOPIC, 1, enemyTeamUpdateCallback);
 
     // Initialize the logger
     Util::Logger::LoggerSingleton::initializeLogger();
