+++ src/thunderbots/software/launch/ai.launch
@@ -1,7 +1,7 @@
 <launch>
 
-    <!-- Launch the backend_input node -->
-    <node name="backend_input" pkg="thunderbots" type="backend_input" output="screen">
+    <!-- Launch the network_input node -->
+    <node name="network_input" pkg="thunderbots" type="network_input" output="screen">
     </node>
 
     <!-- Launch the ai logic node -->
