+++ src/thunderbots/software/network_input/main.cpp
@@ -2,9 +2,9 @@
 
 #include <boost/exception/diagnostic_information.hpp>
 
-#include "backend_input/backend.h"
-#include "backend_input/networking/ssl_vision_client.h"
 #include "geom/point.h"
+#include "network_input/backend.h"
+#include "network_input/networking/ssl_vision_client.h"
 #include "thunderbots_msgs/Ball.h"
 #include "thunderbots_msgs/Field.h"
 #include "thunderbots_msgs/Team.h"
@@ -37,19 +37,19 @@ int main(int argc, char** argv)
     Backend backend = Backend();
 
     // Init ROS node
-    ros::init(argc, argv, "backend_input");
+    ros::init(argc, argv, "network_input");
     ros::NodeHandle node_handle;
 
     // Create publishers
     ros::Publisher ball_publisher = node_handle.advertise<thunderbots_msgs::Ball>(
-        Util::Constants::BACKEND_INPUT_BALL_TOPIC, 1);
+        Util::Constants::NETWORK_INPUT_BALL_TOPIC, 1);
     ros::Publisher field_publisher = node_handle.advertise<thunderbots_msgs::Field>(
-        Util::Constants::BACKEND_INPUT_FIELD_TOPIC, 1);
+        Util::Constants::NETWORK_INPUT_FIELD_TOPIC, 1);
     ros::Publisher friendly_team_publisher =
         node_handle.advertise<thunderbots_msgs::Team>(
-            Util::Constants::BACKEND_INPUT_FRIENDLY_TEAM_TOPIC, 1);
+            Util::Constants::NETWORK_INPUT_FRIENDLY_TEAM_TOPIC, 1);
     ros::Publisher enemy_team_publisher = node_handle.advertise<thunderbots_msgs::Team>(
-        Util::Constants::BACKEND_INPUT_ENEMY_TEAM_TOPIC, 1);
+        Util::Constants::NETWORK_INPUT_ENEMY_TEAM_TOPIC, 1);
 
     // Initialize the logger
     Util::Logger::LoggerSingleton::initializeLogger();
